New charting code.
